move cameraMercatorCoordinateFromCenterAndRotation() to mercator_utils
